# refactor code. # implemented a few more x86 instructions in the visor sim. now understands opcodes: cld, jmp far, mov ptr, movsd,  rep. - the implementation is a bit crude. needs to be moved to xcodeInterp code. # added safe guard to split bios command. fixed bug where split bios was only checking if the romsize is equal and exiting. passing in a bios size less then the romsize resulted in an infinite loop; copying 0byte files to the hdd.~whops.
# bank files (-bank1, -bank2 etc) are inferred for commands that use them. (split bios, combine bios)
for split bios, which also uses the biosfile which is also inferred, you will have to specify the bios file first then trailing -arguments without a switch will be treated as a bank file.
 # added '-simsize' switch for overriding default visor hack sim size. default is 32 bytes.
# added '-d' switch. the dump switch. if specified, commands will dump more things. eg. using with the visor sim command will dump the memory sim to a file for later analysis in a decompiler.
 # added 'bfm key' to bldr keys struct as it's directly above it.
# unifying code.
# testing BFM convertsion code. ~requires a bldr that actually contains the bfm entry code. (retail will loop when you call the bfm entry)
# replaced ulongs with uints. this program expects an int and a long to be 4 bytes,
# fixed encryption flags being wrong when building a bios.
# added cert key, eeprom key switches for patching them in the bldr. (-bld)
# removed decode output from list bios command. you can still use -xcode-decode.

+++ inc/rc4.h
@@ -24,9 +24,9 @@
 
 #include "type_defs.h"
 
-void rc4_key(UCHAR* s, const UCHAR* key, const ULONG len);
-void rc4(UCHAR* s, UCHAR* data, const ULONG len);
+void rc4_key(UCHAR* s, const UCHAR* key, const UINT len);
+void rc4(UCHAR* s, UCHAR* data, const UINT len);
 
-void symmetricEncDec(UCHAR* data, const ULONG len, const UCHAR* key, const ULONG keyLen);
+void symmetricEncDec(UCHAR* data, const UINT len, const UCHAR* key, const UINT keyLen);
 
 #endif // _RC4_H
